`refactor`: revert `toLong` to `toInt` due to require field from gql is `Int!`

+++ features/content/createpost/src/main/java/com/tokopedia/createpost/view/presenter/CreatePostPresenter.kt
@@ -127,7 +127,7 @@ class CreatePostPresenter @Inject constructor(
                     }
             )
         } else {
-            getShopFavoriteStatusUseCase.params = GQLGetShopFavoriteStatusUseCase.createParams(listOf(userSession.shopId.toLong()), "")
+            getShopFavoriteStatusUseCase.params = GQLGetShopFavoriteStatusUseCase.createParams(listOf(userSession.shopId.toInt()), "")
             getShopFavoriteStatusUseCase.execute(
                     {
                         followersCount = it.favoriteData.totalFavorite
